Overloaded insertion operator for transactions/ Moved the header creation to a static GUI class
The bank account class insertion operator overload was a bit messy, so I created the iostream insertion overload for the transactions as well to clean everything up a bit.

Because of that, I moved the header creation to a static GUI class, because the header creation was done in the bank account class and the transactions did not have easy access to those method (and it looks a lot cleaner).