Implemented the builder pattern
I created an Elf and Orc class that are subclasses of the NPC base class.

Next I created a NPCDecorator class that acts as a base class for all the decorator classes. The decorator classes print their attribute to the screen (Farmer, Builder and Shaman).

For easy creation, I make a factory class to create the Elf and Orc classes with all the decorator classes.